requirements.txt
# --- Core Data Processing ---
pyspark==3.5.0
delta-spark==3.1.0
polars==0.20.16


# --- Data Warehouse & Transformation ---
dbt-core==1.7.9
dbt-postgres==1.7.9
psycopg2-binary

# --- Streaming & Ingestion ---
kafka-python
Faker

# --- Web UI & Dashboard ---
streamlit
plotly

# --- High-Performance Connectors ---
connectorx
adbc-driver-sqlite
adbc-driver-postgresql

# --- Utilities ---
python-dotenv
ruff

# ---aws---
boto3>=1.34.0
s3fs>=2024.2.0

pyarrow